backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
pytest==7.4.2
pytest-flask==1.2.0
